refactor: cohere command-r mixin
